Delete is broken

+++ src/main/java/com/incode/todo/services/TaskService.java
@@ -1,17 +1,16 @@
 package com.incode.todo.services;
 
-import com.incode.todo.models.Task;
-import com.incode.todo.models.TaskPatch;
-import com.incode.todo.models.TaskPost;
+import com.incode.todo.models.TaskRequest;
+import com.incode.todo.models.TaskResponse;
 import com.incode.todo.repositories.TaskRepository;
 import com.incode.todo.utils.MapperUtils;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 
 import org.springframework.stereotype.Service;
+import org.springframework.util.MultiValueMap;
 
 import reactor.core.publisher.Mono;
 
@@ -21,30 +20,43 @@ public class TaskService {
 
   private final TaskRepository repository;
 
-  public Mono<List<Task>> getAllTasks() {
+  public Mono<List<TaskResponse>> getAllTasks() {
     return repository.findAll().map(MapperUtils::toModel).collectList();
   }
 
-  public Mono<List<Task>> getTask(String id) {
+  public Mono<List<TaskResponse>> getTask(String id) {
     return repository.findById(UUID.fromString(id)).map(MapperUtils::toListModel);
   }
 
-  public Mono<List<Task>> createTask(TaskPost model) {
+  public Mono<List<TaskResponse>> createTask(TaskRequest model) {
     return repository
       .save(MapperUtils.toEntity(model))
       .map(MapperUtils::toListModel);
   }
 
-  public Mono<List<Task>> updateTask(String id, TaskPatch model) {
+  public Mono<List<TaskResponse>> updateTask(String id, TaskRequest model, MultiValueMap<String, String> params) {
+
+    
     return repository
       .findById(UUID.fromString(id))
-      .filter(Objects::nonNull)
-      .map(entity -> MapperUtils.patchEntity(entity, model))
+      .map(entity -> MapperUtils.patchEntity(entity, model, params))
       .switchIfEmpty(Mono.empty())
       .flatMap(entity -> repository.save(entity).map(MapperUtils::toListModel));
+
   }
 
-  public Mono<Void> removeTask(String id) {
-    return repository.delete(MapperUtils.toEntity(UUID.fromString(id)));
+  public Mono<Void> removeTask(String id,  MultiValueMap<String, String> params) {
+
+    if (MapperUtils.isSoftDelete(params)) {
+      repository
+        .findById(UUID.fromString(id))
+        .map(entity -> MapperUtils.patchSoftDeleteEntity(entity))
+        .switchIfEmpty(Mono.empty())
+        .flatMap(entity -> repository.save(entity));
+    return Mono.empty();
+    } else {
+      return repository.delete(MapperUtils.toEntity(UUID.fromString(id)));
+    }
+    
   }
 }
